pyproject.toml
[tool.poetry]
name = "file-sharing-web"
version = "0.1.0"
description = "Image sharing web application"
authors = ["Group7"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = {extras = ["standart"], version = "^0.27.1"}
alembic = "^1.13.1"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
